fix: 가게 수정 시 메뉴 미포함 (기본 정보만 수정)
